Decouple string & path utilities from dexec #29
Moved the following out of dexec.go into new util package:
 * AddPrefix
 * JoinStringSlices
 * ExtractFileExtension
